refactor: moving footer to its own components folder
- changing all import statement that have footer in it
