+++ src/pages/home/index.tsx
@@ -2,7 +2,7 @@ import Hero from '../../components/layouts/home/hero'
 import About from '../../components/layouts/home/about'
 import Work from '../../components/layouts/home/work'
 import Contact from '../../components/layouts/home/contact'
-import Footer from '../../components/layouts/home/footer'
+import Footer from '../../components/footer/footer'
 
 export default function Homepage() {
   return (
